requirements.txt
simplejson
beautifulsoup4
mechanize
MySQL-python
mysqlclient
pycountry